usb: hcd: Add support for host restart on IAA watchdog timeout
Currenttly when the IAA watchdog is triggered in eHCI,
the USB device is not detected and subsequent removal
and insertion also does not work.

Provide a workaround for this issue by removing and
adding the HCD. Add support for atomic notifier callback
to queue msm_otg_sm_work. This checks if HC has died and
then performs hcd removal and addition.

Change-Id: I1d6b04a8551dc1149c5ee2f129521b80eaa1295e
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>